Changed sample size from 16bit to 32 bit
SInce the 2 least significant bits in the first byte are actually the most significant bits from the sensor,changed the sample size from 16 bit to 32bit

+++ max.h
@@ -27,6 +27,6 @@
 
 
 void MAX_Init(void);
-void MAX_GetFifoSample(uint16_t *data,uint8_t bufferSize);
+void MAX_GetFifoSample(uint32_t *data,uint8_t bufferSize);
 
 #endif
